file : seasonalFestival 분리, 배경색 대비 높임
