+++ src/pages/Main/components/Hero.tsx
@@ -5,6 +5,7 @@ import { Input } from "../../../components/ui/input"
 import { Button } from "../../../components/ui/button"
 import { Badge } from "../../../components/ui/Badge"
 import { motion, AnimatePresence } from "framer-motion"
+import { seasonalFestivals } from "./seasonalFestivals"
 
 
 const floatingKeywords = [
@@ -18,52 +19,7 @@ const floatingKeywords = [
   { text: "청자축제", x: 90, y: 80, delay: 1.8, color: "bg-emerald-100 text-emerald-700" },
 ]
 
-export const seasonalFestivals = [
-  {
-    id: "spring",
-    season: "봄",
-    title: "봄꽃 축제",
-    description: "벚꽃과 함께하는 화사한 봄 축제들",
-    festivals: ["진해 군항제", "여의도 벚꽃축제", "경주 벚꽃축제"],
-    accent: "text-pink-600",
-    color: "bg-gradient-to-br from-pink-50 to-rose-100",
-    border: "border-pink-200 border-2",
-    shadow: "0 0 0 3px #f9a8d4", // 봄: 연핑크 (pink-200)
-  },
-  {
-    id: "summer",
-    season: "여름",
-    title: "여름 물축제",
-    description: "시원한 물과 함께하는 여름 축제들",
-    festivals: ["보령 머드축제", "화천 토마토축제", "무주 반딧불축제"],
-    accent: "text-blue-600",
-    color: "bg-gradient-to-br from-blue-50 to-cyan-100",
-    border: "border-blue-200 border-2",
-    shadow: "0 0 0 3px #bae6fd", // 여름: 연파랑 (cyan-200)
-  },
-  {
-    id: "autumn",
-    season: "가을",
-    title: "가을 단풍축제",
-    description: "아름다운 단풍과 함께하는 가을 축제들",
-    festivals: ["진주 남강 유등축제", "부산 불꽃축제", "안동 탈춤페스티벌"],
-    accent: "text-amber-600",
-    color: "bg-gradient-to-br from-amber-50 to-orange-100",
-    border: "border-amber-200 border-2",
-    shadow: "0 0 0 3px #fde68a", // 가을: 연노랑 (amber-200)
-  },
-  {
-    id: "winter",
-    season: "겨울",
-    title: "겨울 얼음축제",
-    description: "하얀 눈과 함께하는 겨울 축제들",
-    festivals: ["화천 산천어축제", "평창 송어축제", "태백 눈축제"],
-    accent: "text-slate-600",
-    color: "bg-gradient-to-br from-slate-50 to-blue-100",
-    border: "border-slate-200 border-2",
-    shadow: "0 0 0 3px #cbd5e1", // 겨울: 연회색 (slate-200)
-  },
-]
+
 
 interface HeroProps {
   setCurrentSeason: (index: number) => void
@@ -87,7 +43,7 @@ export default function Hero({ setCurrentSeason }: HeroProps) {
       setCurrentSeasonProp(next)
     }, 5000)
     return () => clearInterval(interval)
-  }, [currentSeason])
+  }, [currentSeason, setCurrentSeason])
 
   const handleSearch = (e: React.FormEvent) => {
     e.preventDefault()
@@ -184,7 +140,6 @@ export default function Hero({ setCurrentSeason }: HeroProps) {
     duration: 0.7,
     ease: "easeInOut",
   }}
-  onClick={() => handleFloatingClick(k.text)}
 >
       <div className="relative z-10 w-full max-w-4xl px-4 text-center z-0">
         <h1 className="text-4xl md:text-6xl font-bold text-gray-900 mb-4">
