+++ src/pages/Main/components/seasonalFestivals.ts
@@ -1,0 +1,46 @@
+export const seasonalFestivals = [
+  {
+    id: "spring",
+    season: "봄",
+    title: "봄꽃 축제",
+    description: "벚꽃과 함께하는 화사한 봄 축제들",
+    festivals: ["진해 군항제", "여의도 벚꽃축제", "경주 벚꽃축제"],
+    accent: "text-pink-600",
+    color: "bg-gradient-to-br from-pink-50 to-rose-300",
+    border: "border-pink-200 border-2",
+    shadow: "0 0 0 3px #f9a8d4", // 봄: 연핑크 (pink-200)
+  },
+  {
+    id: "summer",
+    season: "여름",
+    title: "여름 물축제",
+    description: "시원한 물과 함께하는 여름 축제들",
+    festivals: ["보령 머드축제", "화천 토마토축제", "무주 반딧불축제"],
+    accent: "text-blue-600",
+    color: "bg-gradient-to-br from-blue-50 to-cyan-300",
+    border: "border-blue-200 border-2",
+    shadow: "0 0 0 3px #bae6fd", // 여름: 연파랑 (cyan-200)
+  },
+  {
+    id: "autumn",
+    season: "가을",
+    title: "가을 단풍축제",
+    description: "아름다운 단풍과 함께하는 가을 축제들",
+    festivals: ["진주 남강 유등축제", "부산 불꽃축제", "안동 탈춤페스티벌"],
+    accent: "text-amber-600",
+    color: "bg-gradient-to-br from-amber-50 to-orange-300",
+    border: "border-amber-200 border-2",
+    shadow: "0 0 0 3px #fde68a", // 가을: 연노랑 (amber-200)
+  },
+  {
+    id: "winter",
+    season: "겨울",
+    title: "겨울 얼음축제",
+    description: "하얀 눈과 함께하는 겨울 축제들",
+    festivals: ["화천 산천어축제", "평창 송어축제", "태백 눈축제"],
+    accent: "text-slate-600",
+    color: "bg-gradient-to-br from-slate-50 to-blue-300",
+    border: "border-slate-200 border-2",
+    shadow: "0 0 0 3px #cbd5e1", // 겨울: 연회색 (slate-200)
+  },
+]
